Refactor file ID types and update related tests
Modified `FsccMessage.cs` to change `ReparsePointTag` from `ushort` to `uint` and `FileId` from `BigInteger` to `Int128` for consistency and efficiency. Updated assertions in `QueryDirectoryTestCases.cs` to reflect these changes, removing checks for 128-bit file IDs in favor of 64-bit checks. Adjusted logic to ensure alignment with file system capabilities and clarified expected behavior for file IDs across different file systems.

+++ ProtoSDK/MS-FSCC/FsccMessage.cs
@@ -6117,14 +6117,14 @@ namespace Microsoft.Protocols.TestTools.StackSdk.FileAccessService.Fscc
         /// <summary>
         /// MUST be ignored by the receiver. 
         /// </summary>
-        public ushort ReparsePointTag;
+        public uint ReparsePointTag;
 
         /// <summary>
         /// The 128-bit file ID, as specified in section 2.1.10, of the file. 
         /// For file systems that do not support a 128-bit file ID, this field MUST be set to 0, 
         /// and MUST be ignored 
         /// </summary>
-        public BigInteger FileId;
+        public Int128 FileId;
 
         /// <summary>
         /// A sequence of Unicode characters containing the file  name. This field might not be NULL-terminated, and   
@@ -6511,7 +6511,7 @@ namespace Microsoft.Protocols.TestTools.StackSdk.FileAccessService.Fscc
         /// <summary>
         /// MUST be ignored by the receiver. 
         /// </summary>
-        public ushort ReparsePointTag;
+        public uint ReparsePointTag;
 
         /// <summary>
         /// The 64-bit file ID, as specified in section 2.1.10, of the file.
